Review from Evan: pretty_doc,consolidated_notation
