+++ src/pretty_printing/pretty_doc.rs
@@ -2,9 +2,9 @@ use crate::valid_notation::ValidNotation;
 
 /// A node in a "document" that supports the necessary methods to be pretty-printed.
 pub trait PrettyDoc<'d>: Copy {
-    type Id: Eq + Copy;
+    type Id: Eq + Copy + Default;
 
-    /// An id that uniquely identifies this node.
+    /// An id that uniquely identifies this node. It should not be `Id::default()`.
     fn id(self) -> Self::Id;
 
     /// The node's notation.
@@ -15,28 +15,31 @@ pub trait PrettyDoc<'d>: Copy {
     fn num_children(self) -> Option<usize>;
 
     /// Get this node's text, or panic. The pretty printer will only call this method if
-    /// `num_children()` returns `None`. It is safe to panic otherwise.
+    /// `num_children()` returns `None` - it is ok to make this method panic otherwise.
     fn unwrap_text(self) -> &'d str;
 
     /// Get this node's i'th child, or panic. The pretty printer will only call this method if
-    /// `num_children()` returns `Some(n)` for `n > i`. It is safe to panic otherwise.
+    /// `num_children()` returns `Some(n)` for `n > i` - it is ok to make this method panic
+    /// otherwise.
     fn unwrap_child(self, i: usize) -> Self;
 
     /// Get this node's last child, or panic. The pretty printer will only call this method if
-    /// `num_children()` returns `Some(_)`. It is safe to panic otherwise.
+    /// `num_children()` returns `Some(n)` where `n > 0` - it is ok to make this method panic
+    /// otherwise.
+    ///
+    /// This method is redundant with `unwrap_child`, but depending on your document representation
+    /// it could have a more efficient implementation.
     fn unwrap_last_child(self) -> Self;
 
-    // This method exists for efficiency. If the tree is implemented with linked lists,
-    // unwrap_child() could be very slow when there are lots of children. Thus `fold` uses this
-    // method instead.
     /// Access this node's previous sibling, or panic. `parent` is this node's parent, and `i` is
     /// the index of its previous sibling. The pretty printer will only call this method if:
     ///
     /// - `parent.num_children()` returned `Some(n)` for `n > i + 1`, and
     /// - the index of `self` is `i + 1`.
     ///
-    /// It is safe to panic otherwise.
-    fn unwrap_prev_sibling(self, parent: Self, i: usize) -> Self {
-        parent.unwrap_child(i)
-    }
+    /// It is ok to make this method panic otherwise.
+    ///
+    /// This method is redundant with `unwrap_child`, but depending on your document representation
+    /// it could have a more efficient implementation.
+    fn unwrap_prev_sibling(self, parent: Self, i: usize) -> Self;
 }
